backend/requirements.txt

flask==2.0.1
flask-cors==3.0.10
boto3==1.26.137
python-dotenv==0.19.0
gunicorn==20.1.0

